requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
importlib_metadata==8.7.0
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==11.1.1
pyppeteer==2.0.0
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
ruff==0.12.10
sdamgia-api==0.1.7
soupsieve==2.7
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
websockets==10.4
yarl==1.20.1
zipp==3.23.0
